基本完成admin管理

+++ src/main/java/com/tz/mooc/web/CourseController.java
@@ -2,6 +2,7 @@ package com.tz.mooc.web;
 
 import com.tz.mooc.pojo.Course;
 import com.tz.mooc.service.CourseService;
+import com.tz.mooc.service.VideoService;
 import com.tz.mooc.util.Page4Navigator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -12,11 +13,18 @@ import java.util.Optional;
 @RestController
 public class CourseController {
     @Autowired CourseService courseService;
+    @Autowired VideoService videoService;
 
     @GetMapping("/subjects/{sid}/courses")
     public Page4Navigator<Course> list(@PathVariable("sid") int sid, @RequestParam(value = "start", defaultValue = "0") int start, @RequestParam(value = "size", defaultValue = "5") int size) throws Exception {
         start = start<0?0:start;
         Page4Navigator<Course> page =courseService.list(sid, start, size,5);
+        return page;
+    }
+
+    @GetMapping("courses/all")
+    public Page4Navigator<Course> list(@RequestParam(value = "start", defaultValue = "0") int start, @RequestParam(value = "size", defaultValue = "5") int size) throws Exception {
+        Page4Navigator<Course> page =courseService.list(start, size,5);
         return page;
     }
 
@@ -34,8 +42,12 @@ public class CourseController {
 
     @DeleteMapping("/courses/{id}")
     public String delete(@PathVariable("id") int id, HttpServletRequest request)  throws Exception {
-        courseService.delete(id);
-        return null;
+        if(videoService.getByCourse(id).isPresent()){
+            return "-1";
+        }else {
+            courseService.delete(id);
+            return null;
+        }
     }
 
     @PutMapping("/courses/{id}")
